requirements.txt
Flask==3.0.0
requests
gspread
google-auth
gunicorn
